cleanup and split allocator.hpp to separate files
Fixes a bug where the arena's m_end was not updated after reallocating
